pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "meshrun"
version = "0.1.0"
description = "Distributed AI inference on compute you already own."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies (from root pyproject.toml)
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
    "cryptography>=44.0.0",
    "torch>=2.6.0",
    "transformers>=4.48.0",
    # CLI dependencies (from meshrun/pyproject.toml)
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "toml>=0.10.2",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "hypothesis>=6.150.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.15.0",
    "mypy>=1.15.0",
]

[project.scripts]
meshrun = "meshrun.app.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["meshrun*"]
